sherlock_api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.13.0
python-dotenv>=0.19.0
pydantic-settings>=2.0.0
openai>=1.0.0
pinecone-client>=2.2.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.21.0